Get Gatsby to actually build web pages from the docs markdown
This is just the bare-bones rendering of markdown to HTML, with $variable$
expansion.

At this point, I think the biggest thing missing machinery-wise is the
table of contents generated from doc-pages.yml.

Establish a dependency on Node >= 15, since that's when
String.prototype.replaceAll was added, and going through the pain of
writing code without something so basic doesn't seem like a good use of
time.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>